All: App options tree.

+++ MiniloGUI/MiniloGUI 2026/Project_Source/D_SUBTREE_Bit_Loc_AV_P.cpp
@@ -12,7 +12,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 
 	return { 
 		ID::subtree_bit_locations, {}, {
-			ValueTree{ String{ AVP::amp_mod_by_velo_amt },
+			ValueTree{ String{ (int)AVP::amp_mod_by_velo_amt },
 				{
 					{ "bit_0", "byte_37__bit_0" },
 					{ "bit_1", "byte_37__bit_1" },
@@ -24,7 +24,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 				}
 			},
 		// AVP::current_voice_name is plugin-only
-		ValueTree{ String{ AVP::flex_slider_bend_range_neg },
+		ValueTree{ String{ (int)AVP::flex_slider_bend_range_neg },
 				{
 					{ "bit_0", "byte_75__bit_4" },
 					{ "bit_1", "byte_75__bit_5" },
@@ -32,7 +32,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_3", "byte_71__bit_3" },
 				}
 			},
-			ValueTree{ String{ AVP::flex_slider_bend_range_pos },
+			ValueTree{ String{ (int)AVP::flex_slider_bend_range_pos },
 				{
 					{ "bit_0", "byte_75__bit_0" },
 					{ "bit_1", "byte_75__bit_1" },
@@ -40,7 +40,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_3", "byte_75__bit_3" },
 				}
 			},
-			ValueTree{ String{ AVP::flex_slider_range },
+			ValueTree{ String{ (int)AVP::flex_slider_range },
 				{
 					{ "bit_0", "byte_86__bit_0" },
 					{ "bit_1", "byte_86__bit_1" },
@@ -52,7 +52,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_7", "byte_79__bit_6" },
 				}
 			},
-			ValueTree{ String{ AVP::flex_slider_target }, 
+			ValueTree{ String{ (int)AVP::flex_slider_target }, 
 				{
 					{ "bit_0", "byte_82__bit_0" },
 					{ "bit_1", "byte_82__bit_1" },
@@ -63,29 +63,29 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_6", "byte_82__bit_6" },
 				}
 			},
-			ValueTree{ String{ AVP::keyboard_octave },
+			ValueTree{ String{ (int)AVP::keyboard_octave },
 				{
 					{ "bit_0", "byte_83__bit_0" },
 					{ "bit_1", "byte_83__bit_1" },
 					{ "bit_2", "byte_83__bit_2" },
 				}
 			},
-			ValueTree{ String{ AVP::lfo_sync_bpm_on },
+			ValueTree{ String{ (int)AVP::lfo_sync_bpm_on },
 				{
 					{ "bit_0", "byte_78__bit_1" },
 				}
 			},
-			ValueTree{ String{ AVP::lfo_sync_key_on },
+			ValueTree{ String{ (int)AVP::lfo_sync_key_on },
 				{
 					{ "bit_0", "byte_78__bit_0" },
 				}
 			},
-			ValueTree{ String{ AVP::lfo_sync_voice_on },
+			ValueTree{ String{ (int)AVP::lfo_sync_voice_on },
 				{
 					{ "bit_0", "byte_78__bit_02" },
 				}
 			},
-			ValueTree{ String{ AVP::microtune_scale },
+			ValueTree{ String{ (int)AVP::microtune_scale },
 				{
 					{ "bit_0", "byte_89__bit_0" },
 					{ "bit_1", "byte_89__bit_1" },
@@ -97,7 +97,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_7", "byte_87__bit_1" },
 				}
 			},
-			ValueTree{ String{ AVP::pgm_fine_tune },
+			ValueTree{ String{ (int)AVP::pgm_fine_tune },
 				{
 					{ "bit_0", "byte_88__bit_0" },
 					{ "bit_1", "byte_88__bit_1" },
@@ -108,7 +108,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_6", "byte_88__bit_6" },
 				}
 			},
-			ValueTree{ String{ AVP::pgm_level },
+			ValueTree{ String{ (int)AVP::pgm_level },
 				{
 					{ "bit_0", "byte_81__bit_0" },
 					{ "bit_1", "byte_81__bit_1" },
@@ -119,7 +119,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_6", "byte_81__bit_6" },
 				}
 			},
-			ValueTree{ String{ AVP::pgm_transpose },
+			ValueTree{ String{ (int)AVP::pgm_transpose },
 				{
 					{ "bit_0", "byte_92__bit_0" },
 					{ "bit_1", "byte_92__bit_1" },
@@ -128,17 +128,17 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_4", "byte_92__bit_4" },
 				}
 			},
-			ValueTree{ String{ AVP::porta_mode_on },
+			ValueTree{ String{ (int)AVP::porta_mode_on },
 				{
 					{ "bit_0", "byte_78__bit_3" },
 				}
 			},
-			ValueTree{ String{ AVP::porta_sync_bpm_on },
+			ValueTree{ String{ (int)AVP::porta_sync_bpm_on },
 				{
 					{ "bit_0", "byte_78__bit_4" },
 				}
 			},
-			ValueTree{ String{ AVP::porta_time },
+			ValueTree{ String{ (int)AVP::porta_time },
 				{
 					{ "bit_0", "byte_69__bit_0" },
 					{ "bit_1", "byte_69__bit_1" },
@@ -150,7 +150,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_7", "byte_63__bit_5" },
 				}
 			},
-			ValueTree{ String{ AVP::scale_key },
+			ValueTree{ String{ (int)AVP::scale_key },
 				{
 					{ "bit_0", "byte_90__bit_0" },
 					{ "bit_1", "byte_90__bit_1" },
@@ -159,7 +159,7 @@ ValueTree Subtree_Bit_Loc_AV_P::build()
 					{ "bit_4", "byte_90__bit_4" },
 				}
 			},
-			ValueTree{ String{ AVP::voice_mode },
+			ValueTree{ String{ (int)AVP::voice_mode },
 				{
 					{ "bit_0", "byte_73__bit_0" },
 					{ "bit_1", "byte_73__bit_1" },
